requirements.txt
Django==4.2
django-debug-toolbar==4.0.0
django-environ==0.10.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
drf-spectacular==0.26.1